Refactor to remove a layer of abstraction
