+++ src/lib.rs
@@ -1,7 +1,6 @@
-//! openstar-ads is a wrapper around the `ads_client` Rust crate to improve the management of ADS connections and symbol subscriptions.
+//! openstar-ads is a Rust crate to improve the management of ADS connections and symbol subscriptions.
 
 mod ads;
-mod client;
 mod connection;
 pub(crate) mod core;
 mod data_types;
@@ -10,8 +9,6 @@ mod params;
 mod symbol;
 
 pub use self::ads::*;
-pub use self::client::NotificationSubscription;
-pub(crate) use self::client::*;
 pub(crate) use self::connection::*;
 pub use self::core::{
     AdsState, AmsAddr, AmsNetId, AmsPort, DeviceInfo, String, String80, Timeouts, WString,
